requirements.txt
django==5.0.7
djangorestframework
requests
pyyaml
django-cors-headers
 